Basic controller with wheels
