--- Assets/Components/Chassis/Cube.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a6beac926c0c33649958a225639d102a
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
