+++ Assets/Robots/BasicRobot/BasicRobot.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d3d60e81c6d64ed47a58545bbb4199a1
+guid: 0c05ec5840effdb419f389a6b4f94c2d
 PrefabImporter:
   externalObjects: {}
   userData:
