+++ Assets/Robots/BasicRobot/RobotController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b1261dc623183454c883ff1dc0d34929
+guid: 9e91dc2480051034eae359efb3d9bdb0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
