--- Assets/Robots/DeadRobot.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: eab0e583446611d4098873eee55877c0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
