--- Assets/Robots/DeadRobot/BasicRobot_Dumb.prefab
@@ -1,96 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1001 &5633531313853817844
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 1416229793796999382, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_Name
-      value: BasicRobot_Dumb
-      objectReference: {fileID: 0}
-    - target: {fileID: 8328497691871357506, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 1.134
-      objectReference: {fileID: 0}
-    - target: {fileID: 8328497691871357506, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8328497691871357506, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 2.172
-      objectReference: {fileID: 0}
-    - target: {fileID: 8328497691871357506, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8328497691871357506, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8328497691871357506, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: -0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8328497691871357506, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 8328497691871357506, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8328497691871357506, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8328497691871357506, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8328497691871357506, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7819281591530455187, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_Enabled
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 300059451555833461, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_Name
-      value: RobotBase
-      objectReference: {fileID: 0}
-    - target: {fileID: 6073537646520395572, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_Name
-      value: RobotBody
-      objectReference: {fileID: 0}
-    - target: {fileID: 4521698339038788002, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_Name
-      value: BasicRobot_Dumb
-      objectReference: {fileID: 0}
-    - target: {fileID: 4748081522506314215, guid: d3d60e81c6d64ed47a58545bbb4199a1,
-        type: 3}
-      propertyPath: m_Enabled
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: d3d60e81c6d64ed47a58545bbb4199a1, type: 3}
