--- Assets/Robots/DeadRobot/BasicRobot_Dumb.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0c05ec5840effdb419f389a6b4f94c2d
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
